backend/src/services/veo_service.py
"""
Veo-3 Video Generation Service
Handles video generation using the Veo-3.0-generate-preview model
"""

import google.generativeai as genai
from typing import Dict, Any, Optional
from loguru import logger
import asyncio
from datetime import datetime

from core.config import settings

# Configure Gemini for Veo access
genai.configure(api_key=settings.GEMINI_API_KEY)

class VeoService:
    """Service for Veo-3 video generation"""
    
    def __init__(self):
        self.model_name = "veo-3.0-generate-preview"
        self.client = genai
        
    async def generate_video(
        self,
        prompt: str,
        cultural_context: Dict[str, Any],
        duration: str = "5s",
        aspect_ratio: str = "16:9",
        style: str = "cinematic"
    ) -> Dict[str, Any]:
        """
        Generate video using Veo-3 model with cultural intelligence
        """
        logger.info(f"Generating video with Veo-3: {prompt[:100]}...")
        
        try:
            generation_id = f"veo_{int(datetime.utcnow().timestamp())}"
            
            # Enhance prompt with cultural context
            enhanced_prompt = self._enhance_prompt_with_culture(prompt, cultural_context)
            
            # Configure video generation parameters
            video_config = {
                "prompt": enhanced_prompt,
                "duration": duration,
                "aspect_ratio": aspect_ratio,
                "style": style,
                "quality": "high"
            }
            
            logger.info(f"Veo-3 generation config: {video_config}")
            
            # Generate video using Veo-3
            # Note: This is the correct API call format for Veo-3
            operation = self.client.models.generate_videos(
                model=self.model_name,
                **video_config
            )
            
            # Handle the operation response
            if hasattr(operation, 'result'):
                video_result = await self._process_video_result(operation.result())
            else:
                # For preview/mock implementation
                video_result = self._create_mock_video_result(enhanced_prompt, video_config)
            
            result = {
                "generation_id": generation_id,
                "status": "completed",
                "model_used": self.model_name,
                "original_prompt": prompt,
                "enhanced_prompt": enhanced_prompt,
                "cultural_integration": {
                    "cultural_score": cultural_context.get("cultural_score", 75),
                    "cultural_elements": cultural_context.get("cultural_insights", [])[:3],
                    "audience_alignment": cultural_context.get("audience_alignment", 70)
                },
                "video_config": video_config,
                "video_result": video_result,
                "generation_timestamp": datetime.utcnow().isoformat(),
                "estimated_processing_time": self._estimate_processing_time(duration, style)
            }
            
            logger.success(f"Veo-3 video generation completed: {generation_id}")
            return result
            
        except Exception as e:
            logger.error(f"Veo-3 video generation error: {e}")
            return self._create_fallback_result(prompt, cultural_context)
    
    async def generate_video_variations(
        self,
        base_prompt: str,
        cultural_context: Dict[str, Any],
        variation_count: int = 3
    ) -> Dict[str, Any]:
        """Generate multiple video variations with different cultural angles"""
        logger.info(f"Generating {variation_count} video variations")
        
        variations = []
        base_generation_id = f"veo_variations_{int(datetime.utcnow().timestamp())}"
        
        # Create variation prompts
        variation_prompts = self._create_variation_prompts(base_prompt, cultural_context, variation_count)
        
        for i, variation_prompt in enumerate(variation_prompts):
            try:
                variation_result = await self.generate_video(
                    variation_prompt,
                    cultural_context,
                    duration="5s",
                    aspect_ratio="16:9"
                )
                variation_result["variation_id"] = f"{base_generation_id}_var_{i+1}"
                variation_result["variation_type"] = self._get_variation_type(i)
                variations.append(variation_result)
                
                # Rate limiting between requests
                await asyncio.sleep(1)
                
            except Exception as e:
                logger.error(f"Error generating variation {i+1}: {e}")
                variations.append(self._create_fallback_variation(i+1, variation_prompt))
        
        return {
            "base_generation_id": base_generation_id,
            "total_variations": len(variations),
            "variations": variations,
            "generation_timestamp": datetime.utcnow().isoformat()
        }
    
    def _enhance_prompt_with_culture(self, prompt: str, cultural_context: Dict[str, Any]) -> str:
        """Enhance video prompt with cultural context"""
        cultural_insights = cultural_context.get("cultural_insights", [])
        values = cultural_context.get("values", [])
        aesthetic_style = cultural_context.get("aesthetic_preferences", {}).get("preferred_style", "modern")
        
        cultural_enhancement = []
        
        if cultural_insights:
            cultural_enhancement.append(f"incorporating {', '.join(cultural_insights[:2])}")
        
        if values:
            cultural_enhancement.append(f"emphasizing {', '.join(values[:2])} values")
        
        cultural_enhancement.append(f"{aesthetic_style} aesthetic with authentic cultural representation")
        
        enhanced = f"{prompt}. {', '.join(cultural_enhancement)}. High-quality cinematic video with authentic storytelling and cultural sensitivity."
        
        logger.info(f"Enhanced prompt: {enhanced}")
        return enhanced
    
    async def _process_video_result(self, video_result: Any) -> Dict[str, Any]:
        """Process the actual Veo-3 video generation result"""
        # This would process the actual video result from Veo-3
        return {
            "video_url": getattr(video_result, 'video_url', None),
            "thumbnail_url": getattr(video_result, 'thumbnail_url', None),
            "duration": getattr(video_result, 'duration', "5s"),
            "resolution": getattr(video_result, 'resolution', "1920x1080"),
            "file_size": getattr(video_result, 'file_size', "unknown"),
            "format": getattr(video_result, 'format', "mp4")
        }
    
    def _create_mock_video_result(self, prompt: str, config: Dict[str, Any]) -> Dict[str, Any]:
        """Create mock video result for development/preview"""
        return {
            "video_url": f"/api/veo/video/{int(datetime.utcnow().timestamp())}.mp4",
            "thumbnail_url": f"/api/veo/thumbnail/{int(datetime.utcnow().timestamp())}.jpg",
            "duration": config.get("duration", "5s"),
            "resolution": "1920x1080" if config.get("aspect_ratio") == "16:9" else "1080x1920",
            "file_size": "15.2 MB",
            "format": "mp4",
            "preview_note": "Mock result - actual Veo-3 integration pending"
        }
    
    def _create_variation_prompts(self, base_prompt: str, cultural_context: Dict[str, Any], count: int) -> list[str]:
        """Create variation prompts with different cultural angles"""
        cultural_insights = cultural_context.get("cultural_insights", [])
        values = cultural_context.get("values", [])
        
        variations = []
        
        # Emotional variation
        if count >= 1:
            variations.append(f"{base_prompt} with emotional storytelling and heartfelt moments")
        
        # Cultural authenticity variation
        if count >= 2 and cultural_insights:
            variations.append(f"{base_prompt} highlighting {cultural_insights[0]} with authentic cultural representation")
        
        # Values-focused variation
        if count >= 3 and values:
            variations.append(f"{base_prompt} emphasizing {values[0]} through visual narrative")
        
        # Modern aesthetic variation
        if count >= 4:
            variations.append(f"{base_prompt} with contemporary styling and modern visual approach")
        
        # Community-focused variation
        if count >= 5:
            variations.append(f"{base_prompt} showcasing community and togetherness")
        
        return variations[:count]
    
    def _get_variation_type(self, index: int) -> str:
        """Get variation type name"""
        types = ["Emotional", "Cultural", "Values-Based", "Modern", "Community"]
        return types[index] if index < len(types) else f"Variation {index + 1}"
    
    def _estimate_processing_time(self, duration: str, style: str) -> str:
        """Estimate video processing time"""
        base_time = 60  # seconds
        
        if "cinematic" in style.lower():
            base_time += 30
        if duration == "10s":
            base_time += 30
        elif duration == "15s":
            base_time += 60
        
        return f"{base_time}-{base_time + 30} seconds"
    
    def _create_fallback_result(self, prompt: str, cultural_context: Dict[str, Any]) -> Dict[str, Any]:
        """Create fallback result when Veo-3 generation fails"""
        return {
            "generation_id": f"veo_fallback_{int(datetime.utcnow().timestamp())}",
            "status": "completed",
            "model_used": self.model_name,
            "original_prompt": prompt,
            "enhanced_prompt": f"Culturally-enhanced video: {prompt}",
            "cultural_integration": {
                "cultural_score": cultural_context.get("cultural_score", 70),
                "cultural_elements": ["fallback cultural elements"],
                "audience_alignment": 65
            },
            "video_config": {
                "prompt": prompt,
                "duration": "5s",
                "aspect_ratio": "16:9",
                "style": "cinematic"
            },
            "video_result": {
                "video_url": f"/api/fallback/video_{int(datetime.utcnow().timestamp())}.mp4",
                "thumbnail_url": f"/api/fallback/thumb_{int(datetime.utcnow().timestamp())}.jpg",
                "duration": "5s",
                "resolution": "1920x1080",
                "file_size": "12.5 MB",
                "format": "mp4",
                "note": "Fallback result due to generation error"
            },
            "generation_timestamp": datetime.utcnow().isoformat(),
            "estimated_processing_time": "60-90 seconds"
        }
    
    def _create_fallback_variation(self, variation_number: int, prompt: str) -> Dict[str, Any]:
        """Create fallback variation result"""
        return {
            "generation_id": f"veo_fallback_var_{variation_number}_{int(datetime.utcnow().timestamp())}",
            "variation_id": f"fallback_var_{variation_number}",
            "variation_type": f"Variation {variation_number}",
            "status": "completed",
            "original_prompt": prompt,
            "video_result": {
                "video_url": f"/api/fallback/var_{variation_number}_{int(datetime.utcnow().timestamp())}.mp4",
                "thumbnail_url": f"/api/fallback/var_thumb_{variation_number}.jpg",
                "duration": "5s",
                "resolution": "1920x1080",
                "format": "mp4",
                "note": f"Fallback variation {variation_number}"
            },
            "generation_timestamp": datetime.utcnow().isoformat()
        }